+++ R/Learner.R
@@ -550,6 +550,10 @@ get_log_condition = function(state, condition) {
   }
 }
 
+#' @export
+default_values.Learner = function(x, search_space, task, ...) { # nolint
+  default_values(x$param_set)[search_space$ids()]
+}
 # #' @export
 # format_list_item.Learner = function(x, ...) { # nolint
 #   sprintf("<lrn:%s>", x$id)